utils 추가 및 기존 페이지 refactor
