+++ src/hoc/auth.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { useEffect } from "react";
-import { NEED_TO_LOGIN } from "../constants/ERROR_MESSAGE";
 import { authService } from "../firebase";
+import { NEED_TO_LOGIN } from "../utility/ALERT_MESSAGE";
 
 export default function (SpecificComponent) {
   function AuthCheck(props) {
