+++ src/pages/CourseRegisterPage/CourseRegitserBox/CourseUpdate.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
-import { SUCCESS_APPLICATION } from "../../../constants/ERROR_MESSAGE";
 import { firestoreService } from "../../../firebase";
+import { SUCCESS_APPLICATION } from "../../../utility/ALERT_MESSAGE";
 import CourseRegisterBox from "./CourseRegisterBox";
 
 export default function CourseUpdate() {
