--- src/style/Colors.jsx
@@ -1,2 +1,0 @@
-export const baseColor = "rgb(245, 245, 245)";
-export const mainColor = "#c32020";
