+++ src/utility/COLORS.js
@@ -1,0 +1,2 @@
+export const BASE_COLOR = "rgb(245, 245, 245)";
+export const MAIN_COLOR = "#c32020";
