+++ src/utility/COMMON_FUNCTION.js
@@ -1,6 +1,6 @@
-export default function () {
+export const RandomEmoji = () => {
   return emojis[Math.floor(Math.random() * emojis.length)];
-}
+};
 
 const emojis = [
   "🧕",
